% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agricultural.land.consumption.R
\name{agricultural_land_consumption}
\alias{agricultural_land_consumption}
\title{This function calculates the amount of high-value agricultural land that is
estimated
to become urban settlements, between the base year and the horizon year.}
\usage{
agricultural_land_consumption(agr_land_base, agr_land_horizon)
}
\arguments{
\item{agr_land_base}{a raster that contains agricultural land base information}

\item{agr_land_horizon}{a raster that contains buildup informationt in the base year}
}
\value{
a data frame with information about the losses in agricultural land
in square kilometers
}
\description{
METHOD:
Agricultural land consumption (agricultural.consumption) is calculated as the
area of the
urban footprint (buildup) that was land with high agricultural value in the
base year.
The first step is to define the polygon or the area that acknowledges
high-value agricultural land.
The agricultural land lost to urbanization (agri_c) is calculated by adding
up the hectares of urban area (buildup)
located within the agricultural polygon (agri_b).
}
\examples{
library(raster)

land.cover2025 <- system.file("extdata", "Land_cover.tif",
  package =
    "urbanperformance"
)
land.cover2025 <- raster::raster(land.cover2025)
footprint.2030 <- system.file("extdata", "Build_up_2030.tif",
  package =
    "urbanperformance"
)
footprint.2030 <- raster::raster(footprint.2030)
agri.land <- land.cover2025
agri.land[agri.land != 1] <- 0

agri.consumption <- agricultural_land_consumption(agri.land, footprint.2030)
}
