% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amenities.checker.R
\name{amenities.checker}
\alias{amenities.checker}
\title{This function adjust the vector layer that cointains information about urban
amenities,
standardize the CRS and ensure the points and polygons are inside the area of
interest}
\usage{
amenities.checker(..., aoi, proj = 4326)
}
\arguments{
\item{...}{a shapefile or list of shapefile that contains polygons or points
with the localition of urban amenities.
The shapefile must contain the "fclass" column.}

\item{aoi}{a shapefile that contains the delimitation of the study area}

\item{proj}{CRS coordinates}
}
\value{
a standardize shapefile that contains the amenities points that are
inside the study area.
}
\description{
This function adjust the vector layer that cointains information about urban
amenities,
standardize the CRS and ensure the points and polygons are inside the area of
interest
}
\examples{
library(sf)
data(amenities.cun)
data(aoi.cun)


amenities <- amenities.checker(amenities.cun, aoi = aoi.cun)
}
