% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amenities.cun}
\alias{amenities.cun}
\title{Amenities data}
\format{
A spatial points data frame.
}
\source{
Simulated data for the urbanperformance package.
}
\usage{
amenities.cun
}
\description{
A dataset containing amenities data.
}
\keyword{datasets}
