% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amenities.proximity.R
\name{amenities.proximity}
\alias{amenities.proximity}
\title{This function calculates the total and the percentage of the population that
lives within the maximum recommended distance to each of the following urban
public services and amenities: schools, health facilities (clinics and
hospitals),
cultural facilities (community centers, libraries, theaters),
worship places, markets, sport facilities, public spaces (parks, green areas,
squares, playgrounds),
and any other provided.}
\usage{
amenities.proximity(..., pop, parameters = NULL, save = TRUE)
}
\arguments{
\item{...}{a shapefile or list of shapefiles that contains polygons or points
with the location of urban amenities.
The shapefile must contain the "fclass" column.}

\item{pop}{raster with information about population distribution}

\item{parameters}{table with the maximum distance values for each fclass.
Base variable is null, and the indicator use the
p.distances table included in the package that contains general standards
related to the maximum distance recommended for each
fclass to calculate the proximities.}

\item{save}{specify if the raster distance will be saved in the global
environment}
}
\value{
a data frame with the results of the indicator with
the total and percentage of population with proximity to urban amenities
}
\description{
METHOD:
One indicator is calculated for each of the provided classes of amenities.
The proximity (proxi) is calculated for each amenity class (fclassi) by
dividing the
population (pop_prox_ami) that lives within the maximum distance recommended
for that type of amenity (max_disti),
by the total population (tot_pop).
}
\details{
The first step is to calculate a raster that contains the linear distance
from the center of each amenity.
Next, a selection of pixels that contain values lower or equal to the maximum
recommended distance for each amenity.
Maximum recommended distance values for base amenities can be found in the
‘p.distances’ table of this package.
Next, the population (pop) of all the pixels contained in the raster
selection is added up to obtain the population that
has access to a particular amenity (proxi). Finally, this population is
divided by the total population of the
urban area (tot.pop) to obtain the percentage of the population that lives
within the recommended distance for that
type of amenity (r.proximity).
}
\examples{
library(sf)
library(raster)
data(amenities.cun)

pop.b <- system.file("extdata", "POP_2025.tif", package = "UPtooltest")
pop.b <- raster::raster(pop.b)

amen.proximity <- amenities.proximity(amenities.cun, pop = pop.b)
}
