% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aoi.cun}
\alias{aoi.cun}
\title{AOI data}
\format{
A spatial polygon data frame.
}
\source{
Simulated data for the urbanperformance package.
}
\usage{
aoi.cun
}
\description{
A dataset containing the Area of Interest (AOI).
}
\keyword{datasets}
