% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle.cun}
\alias{cycle.cun}
\title{Cycle data}
\format{
A spatial lines data frame.
}
\source{
Simulated data for the urbanperformance package.
}
\usage{
cycle.cun
}
\description{
A dataset containing cycle tracks.
}
\keyword{datasets}
