% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.track.density.R
\name{cycle_track_density}
\alias{cycle_track_density}
\title{This function calculates the cycle track kilometers per square kilometer of
urban area.}
\usage{
cycle_track_density(cycle, footprint)
}
\arguments{
\item{cycle}{shapefile that contains cycle tracks}

\item{footprint}{raster that contains the urban footprint in the base year}
}
\value{
a data frame with information about the total km of cycle tracks per
square kilometer
}
\description{
METHOD:
The cycling infrastructure coverage (cycle.dens) is calculated by adding up
the
length of each cycle path (c.lengthm) in the urban area and dividing this
total (c.length) by
the urban area’s footprint (u.footprint).
}
\examples{
library(sf)
library(raster)
data(cycle.cun)

footprint.b <- system.file("extdata", "Build_up_2025.tif",
  package =
    "UPtooltest"
)
footprint.b <- raster::raster(footprint.b)

cycle.dens <- cycle_track_density(cycle.cun, footprint.b)
}
