% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green.land.consumption.R
\name{green.land.consumption}
\alias{green.land.consumption}
\title{This function calculates the amount of valuable natural land that is
estimated
to become urban settlements, between the base year and the horizon year.}
\usage{
green.land.consumption(greenl, buildup)
}
\arguments{
\item{greenl}{polygon or raster that contains the area of valuable green
land}

\item{buildup}{raster that contains the urban footprint in the base year}
}
\value{
a data frame with information about the losses in green land in
square kilometers
}
\description{
METHOD:
Green land consumption (greenl.consumption) is calculated as the area of the
urban footprint (buildup) that was land with valuable green land in the base
year.
The first step is to define the polygon or the area that acknowledges
valueable land.
The green land lost to urbanization (greenl.c) is calculated by adding up the
hectares of urban area (buildup)
located within the green land polygon (greenl.b).
}
\examples{
library(raster)

land.cover2025 <- system.file("extdata", "Land_cover.tif", package =
"UPtooltest")
land.cover2025 <- raster::raster(land.cover2025)
footprint.2030 <- system.file("extdata", "Build_up_2030.tif", package =
"UPtooltest")
footprint.2030 <- raster::raster(footprint.2030)
green.land <- land.cover2025
green.land[green.land != 8] <- 0

green.consumption <- green.land.consumption(green.land, footprint.2030)
}
