% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green.area.pcapita.R
\name{green_area_pcapita}
\alias{green_area_pcapita}
\title{This function calculates the total green area within the urban area
(forests, parks, gardens, etc.) per inhabitant.}
\usage{
green_area_pcapita(ga, pop)
}
\arguments{
\item{ga}{raster layer or shapefile that contains the polygons for green
areas}

\item{pop}{raster of population distribution or numeric value with
information of total population}
}
\value{
a data frame with information about the infill areas in square
kilometers
}
\description{
METHOD:
The park and public space area available per person (green_area.p) is
estimated
by adding up the area with vegetation (green_area), in square meters, in the
parks
and public open spaces of the urban area, and dividing this result
(green_area.t) by the total population (t.pop).
}
\examples{
library(raster)

land.cover2025 <- system.file("extdata", "Land_cover.tif",
  package =
    "urbanperformance"
)
land.cover2025 <- raster::raster(land.cover2025)
pop_base <- system.file("extdata", "POP_2025.tif",
  package = "urbanperformance"
)
pop_base <- raster::raster(pop_base)

green.land <- land.cover2025
green.land[green.land != 8] <- 0
green_a_capita <- green_area_pcapita(green.land, pop_base)
}
