% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.in.risk.R
\name{hazard.exposure}
\alias{hazard.exposure}
\title{This function calculates the number and percentage of the population that is
exposed to hazards because they live within the zone of influence
of natural and anthropogenic sources.}
\usage{
hazard.exposure(r, pop)
}
\arguments{
\item{r}{a raster or raster stack with delimitation of risk areas}

\item{pop}{a raster o raster stack with information about population
distribution}
}
\value{
a data frame with information about total population in inhabitants
and percentage
}
\description{
METHOD:
Exposure (hazard.exposure) is calculated by dividing the number of
inhabitants (r.pop)
that live within the zone of influence of a hazard zone by the total
population (p).
First, clean raster process in apply to convert the layer into a binary
raster. Each type of
hazard is identified by its layer name. Second, the population (r.pop) of all
the analysis
points contained in the hazard area is added up to obtain the population that
lives within the affectation zone (r.pop).
Finally, the population is divided by the total population of the urban area
(p) to obtain the exposure
(hazard.exposure) expressed as a percentage.
}
\examples{
library(raster)

pop.b <- system.file("extdata", "POP_2025.tif", package = "UPtooltest")
pop.b <- raster::raster(pop.b)

hazard <- system.file("extdata", "Hazard.tif", package = "UPtooltest")
hazard <- raster::raster(hazard)

pop.in.hazard <- hazard.exposure(hazard, pop = pop.b)
}
