% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.consumption.R
\name{land.consumption}
\alias{land.consumption}
\title{This function calculates land consumption between two urban growth scenarios.
Land consumption is the amount of land predicted to change from natural
habitats or
agricultural uses into urban human settlements between the base year and the
horizon year.}
\usage{
land.consumption(fp.base, fp.horizon)
}
\arguments{
\item{fp.base}{a raster that contains the footprint in the base year}

\item{fp.horizon}{a raster that contains the footprint in the horizon year}
}
\value{
a data frame with information about the land consumption indicator
calculated in square kilometers
}
\description{
METHOD:
Land consumption (land.consumption) is calculated as the difference between
the city footprint in the horizon year (fp.horizon) and the footprint in the
base
year (fp.base). The city footprint refers to the total built-up area of a
city,
including streets, open space, and inner vacant land.
}
\examples{
library(raster)

footprint.base <- system.file("extdata", "Build_up_2025.tif", package =
"UPtooltest")
footprint.base <- raster::raster(footprint.base)
footprint.2030 <- system.file("extdata", "Build_up_2030.tif", package =
"UPtooltest")
footprint.2030 <- raster::raster(footprint.2030)

l.consumption <- land.consumption(footprint.base, footprint.2030)
}
