% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land.cover.classes}
\alias{land.cover.classes}
\title{Land cover classes}
\format{
A data frame with columns:
\describe{
  \item{cat}{Integer category identifier}
  \item{fclass}{Land cover or urban land-use classification description}
}
}
\source{
Simulated data and classification scheme for the urbanperformance package.
}
\usage{
land.cover.classes
}
\description{
A dataset containing regional land cover and urban zoning classifications.
}
\keyword{datasets}
