% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.cover.loss.R
\name{land_cover_loss}
\alias{land_cover_loss}
\title{This function calculates the amount of a specific land coverage that is
become
to a different land cover, between different scenarios. This indicator can be
calculate for
a specific land coverage loss or different coverages including in the same
raster.}
\usage{
land_cover_loss(..., areas = NULL)
}
\arguments{
\item{...}{raster layers or a raster stack that contains land cover areas}

\item{areas}{base parameter is NULL, user can load a data frame with the land
categories}
}
\value{
a data frame with information about the losses in land coverage in
square kilometers
}
\description{
METHOD:
Land cover loss (land_cover_loss) is calculated as the difference in areas of
a specific land cover in two different periods.
First the area of each coverage is calculated for the base year using the
function land_cover_areas included in the
package. Next, the areas of each coverage is calculated for the horizon year
or for the scenario or period selected.
Finally, the difference between both scenarios is calculated (loss).
}
\examples{
library(raster)
library(reshape2)

land.cover2018 <- system.file("extdata", "Land_cover_2018.tif",
  package =
    "urbanperformance"
)
land.cover2018 <- raster::raster(land.cover2018)
land.cover2025 <- system.file("extdata", "Land_cover.tif",
  package =
    "urbanperformance"
)
land.cover2025 <- raster::raster(land.cover2025)

land.cover <- raster::stack(land.cover2018, land.cover2025)

land_loss <- land_cover_loss(land.cover)
}
