% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.project.R
\name{load_project_state}
\alias{load_project_state}
\title{Load Project State from Directory or .up File}
\usage{
load_project_state(project)
}
\arguments{
\item{project}{Character path to project directory or `.up` file.}
}
\value{
A list containing the deserialized project state.
}
\description{
Load Project State from Directory or .up File
}
