% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non.cero.R
\name{non_cero}
\alias{non_cero}
\title{This function change the values in the raster to convert values higher than 0
into 1 and NA values into 0}
\usage{
non_cero(r)
}
\arguments{
\item{r}{a raster layer with values >= 0}
}
\value{
a binary raster with values 1 and 0
}
\description{
This function change the values in the raster to convert values higher than 0
into 1 and NA values into 0
}
