% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.cero.R
\name{one.cero}
\alias{one.cero}
\title{This function change the values in the raster to convert values higher than 0
into 1 and NA values into 0}
\usage{
one.cero(r)
}
\arguments{
\item{r}{a raster layer with values >= 0}
}
\value{
a binary raster with values 1 and 0
}
\description{
This function change the values in the raster to convert values higher than 0
into 1 and NA values into 0
}
