% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.density.R
\name{population_density}
\alias{population_density}
\title{This function calculates population density in a scenario.
Population density is the amount of inhabitants per built-up area
in a scenario}
\usage{
population_density(pop, fp)
}
\arguments{
\item{pop}{a raster layer or raster stack that contains the population
distribution in the base or horizon year}

\item{fp}{a raster layer or raster stack that contains the footprint in the
base or horizon year}
}
\value{
a data frame with information about the population density indicator
calculated in inhabitants per square kilometer,
total population, and the total footprint calculated in square kilometers per
each scenario.
}
\description{
METHOD:
The population density (population density) is calculated by dividing the
total number of inhabitants
(pop_tot) by the total built-up area (urban_fp).
The total population of the urban area (pop_tot) is calculated as the sum
of the population of each pixel, this calculation uses the total population
function (tot_pop), while the
total build-up area is calculated as the sum of build-up area of each pixel,
using the urban footprint function (urban_footprint)
The total population and urban footprint functions are included in this
package.
NOTE: If user provide a stack in any of the parameters, the should must have
the same number of layers for population
and footprint stack.
}
\examples{
library(raster)
population.b <- system.file("extdata", "POP_2025.tif",
  package =
    "urbanperformance"
)
population.b <- raster::raster(population.b)
footprint.base <- system.file("extdata", "Build_up_2025.tif",
  package =
    "urbanperformance"
)
footprint.base <- raster::raster(footprint.base)

pop_dens <- population_density(population.b, footprint.base)
}
