% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.transport.proximity.R
\name{public_transport_proximity}
\alias{public_transport_proximity}
\title{This function calculates the total and the percentage of the population that
lives within a maximum recommended distance from a public transport station,
according
to the transport classification.}
\usage{
public_transport_proximity(..., pop, parameters = NULL)
}
\arguments{
\item{...}{a shapefile or list of shapefiles that contains polygons or points
with the location of public
transportation stops or lines. The shapefile must contain the "fclass"
column.}

\item{pop}{raster with information about population distribution}

\item{parameters}{base variable is null, but the indicator can use the
p.distances tables that contains information about the maximum distance
recommended for each
fclass to calculate the proximities}
}
\value{
a data frame with the results of the indicator with
the total and percentage population with proximity to transport stops or
services
}
\description{
METHOD:
Public transport proximity (transport.proximity) is calculated by dividing
the population
(pop_prox_transit) that lives within the maximum recommended distance to a
public transport route or stop,
by the total population (pop).
First, a raster that contains the linear distance from bus stops location is
calculated (dist_r),
according to the type of transportation system (fclass). The maximum
recommended distance varies according
to the type of transportation.
Second, the population (pop) of all the analysis points contained in the
buffer is added up
to obtain the population that lives close to public transport
(pop_prox_transit).
Third, this population is divided by the total population of the urban area
(tot_pop) to
obtain the percentage of the population that lives close to public transport
(transport.proximity).
}
\examples{
library(sf)
library(raster)
data(transport.cun)

pop_base <- system.file("extdata", "POP_2025.tif",
  package = "urbanperformance"
)
pop_base <- raster::raster(pop_base)

pt_prox <- public_transport_proximity(transport.cun, pop = pop_base)
}
