% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterchecker.R
\name{rasterchecker}
\alias{rasterchecker}
\title{This function read and resample the rasters and convert all the raster data
into a stack}
\usage{
rasterchecker(..., base)
}
\arguments{
\item{...}{one or more than two rasters or a raster stack.}

\item{base}{base layer that contains the extent and resolution desired.}
}
\value{
A standardize raster, or rasters with the same extent and resolution,
and projection
}
\description{
This function read and resample the rasters and convert all the raster data
into a stack
}
\examples{
library(raster)
library(raster)

pop.b <- system.file("extdata", "POP_2025.tif", package = "UPtooltest")
pop.b <- raster::raster(pop.b)

pop.2030 <- system.file("extdata", "POP_2030.tif", package = "UPtooltest")
pop.2030 <- raster::raster(pop.2030)

pop.corrected <- rasterchecker(pop.2030, base = pop.b)
}
