% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roads.cun}
\alias{roads.cun}
\title{Roads data}
\format{
A spatial lines data frame.
}
\source{
Simulated data for the urbanperformance package.
}
\usage{
roads.cun
}
\description{
A dataset containing roads.
}
\keyword{datasets}
