% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roads.density.R
\name{roads.density}
\alias{roads.density}
\title{This function calculates the roads kilometers per square kilometer of urban
area.}
\usage{
roads.density(roads, footprint)
}
\arguments{
\item{roads}{shapefile that contains  roads}

\item{footprint}{raster that contains the urban footprint in the base year}
}
\value{
a data frame with information about the total km of roads per square
kilometer
}
\description{
METHOD:
The road density in the urban area (roads.dens) is calculated by adding up
the
length of each road (r.lengthm) in the urban area and dividing this total
(r.length) by
the urban area’s footprint (u.footprint).
}
\examples{
library(sf)
library(raster)
data(roads.cun)

footprint.b <- system.file("extdata", "Build_up_2025.tif",
  package =
    "UPtooltest"
)
footprint.b <- raster::raster(footprint.b)

roads.d <- roads.density(roads.cun, footprint.b)
}
