% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roads.length.R
\name{roads_length}
\alias{roads_length}
\title{This function calculates the roads kilometers per each fclass in the study
area}
\usage{
roads_length(roads)
}
\arguments{
\item{roads}{shapefile that contains roads. This layer must contain the
"fclass" colum}
}
\value{
a data frame with information about the total km of roads of each
category in kilometers
}
\description{
METHOD:
The roads length (roads_length) is calculated by adding up the
length of road roads in kilometers (lengthkm) for each category (fclass)
included.
}
\examples{
library(sf)
data(roads.cun)

r_length <- roads_length(roads.cun)
}
