% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.project.R
\name{save_project_state}
\alias{save_project_state}
\title{Save Project State to Directory and .up File}
\usage{
save_project_state(project_dir, state)
}
\arguments{
\item{project_dir}{Character. Root path of the project.}

\item{state}{List. Reactive state object to serialize.}
}
\value{
Normalized path to saved metadata file.
}
\description{
Save Project State to Directory and .up File
}
