% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot.pop.R
\name{tot_pop}
\alias{tot_pop}
\title{This function calculates the total population of a specific scenario or year}
\usage{
tot_pop(ras)
}
\arguments{
\item{ras}{a raster that contains population information}
}
\value{
a data frame with information about the total population in
inhabitants
}
\description{
This function calculates the total population of a specific scenario or year
}
\examples{
library(raster)

pop_base <- system.file("extdata", "POP_2025.tif", package = "urbanperformance")
pop_base <- raster::raster(pop_base)

total_population <- tot_pop(pop_base)
}
