% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transport.cun}
\alias{transport.cun}
\title{Transport data}
\format{
A spatial lines data frame.
}
\source{
Simulated data for the urbanperformance package.
}
\usage{
transport.cun
}
\description{
A dataset containing public transport networks.
}
\keyword{datasets}
