% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urba.footprint.R
\name{urban.footprint}
\alias{urban.footprint}
\title{This function calculates the urban footprint for a specific scenario/year}
\usage{
urban.footprint(ras)
}
\arguments{
\item{ras}{one raster that contains buildup information}
}
\value{
a data frame with information about the urban footprint area in
square kilometers
}
\description{
METHOD:
Urban footprint (urban.footprint) of the urban area in a specific year or
scenario is calculated
by adding up the area of each pixel with buildup (r), then divide it by 1e6
to convert the value to square kilometer.
}
\examples{
library(raster)

footprint.b <- system.file("extdata", "Build_up_2025.tif",
  package =
    "UPtooltest"
)
footprint.b <- raster::raster(footprint.b)

u.footprint25 <- urban.footprint(footprint.b)
}
