% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urba.footprint.R
\name{urban_footprint}
\alias{urban_footprint}
\title{This function calculates the urban footprint for a specific scenario/year}
\usage{
urban_footprint(ras)
}
\arguments{
\item{ras}{one raster that contains buildup information}
}
\value{
a data frame with information about the urban footprint area in
square kilometers
}
\description{
METHOD:
Urban footprint (urban_footprint) of the urban area in a specific year or
scenario is calculated
by adding up the area of each pixel with buildup (r), then divide it by 1e6
to convert the value to square kilometer.
}
\examples{
library(raster)

footprint_b <- system.file("extdata", "Build_up_2025.tif",
  package = "urbanperformance"
)
footprint_b <- raster::raster(footprint_b)

u_footprint25 <- urban_footprint(footprint_b)
}
