% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbanperformance-package.R
\docType{package}
\name{urbanperformance-package}
\alias{urbanperformance}
\alias{urbanperformance-package}
\title{urbanperformance: Urban Performance}
\description{
This package includes calculation methods to estimate Urban Performance indicators. It provides functions to analyze spatial geometries and calculate variables over specific zones.
}
\author{
\strong{Maintainer}: Ricardo Ochoa \email{ricardo.ochoa@capsus.mx}

Authors:
\itemize{
  \item Socorro Roman [contributor]
}

}
\keyword{internal}
