% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land.cover.classes}
\alias{land.cover.classes}
\title{Land cover classes}
\format{
A data frame with land cover classes and descriptions.
}
\source{
Simulated data for the urbanperformance package.
}
\usage{
land.cover.classes
}
\description{
A dataset containing land cover classifications.
}
\keyword{datasets}
