% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p.distances}
\alias{p.distances}
\title{Proximity distances}
\format{
A numeric vector.
}
\source{
Simulated data for the UPtooltest package.
}
\usage{
p.distances
}
\description{
A dataset containing proximity distance thresholds.
}
\keyword{datasets}
