% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPtooltest-package.R,
%   R/urbanperformance-package.R
\docType{package}
\name{urbanperformance-package}
\alias{urbanperformance-package}
\title{urbanperformance: UP indicators package}
\description{
This package includes calculation methods to estimate Urban Performance indicators. It provides functions to analyze spatial geometries and calculate variables over specific zones.

This package includes calculation methods to estimate Urban Performance indicators. It provides functions to analyze spatial geometries and calculate variables over specific zones.
}
\keyword{internal}
